Restructured Pearson correlation testing directory
To facilitate simultaneous and focused programming, the following
changes have been made to the Pearson correlation testing directory:

- Now there's a "header" file, called common.jl, which contains functions
  and package imports that will be needed in all cases.
- For now, each case has its own file. The first line in these files
  should be:
    include("common.jl")
- Tests for output should be contained in separate files, possibly
  following the scheme test_[name].jl. The first line in these files
  should be e.g.,
    include("pc_cont_cont.jl")
  depending on what case is being tested.
