+++ scratch/pearson_matching_Xioa_Zhou/pc_cont_cont.jl
@@ -1,49 +1,4 @@
-import Statistics: mean, std, quantile
-import FastGaussQuadrature: gausshermite
-import Polynomials: Polynomial
-import IntervalRootFinding: roots
-import Memoize: @memoize
-
-using Distributions
-using IntervalArithmetic
-
-
-margins = [
-    (Beta(2, 3)),
-    (Normal(12, 2)),
-    (Exponential(3.14))
-]
-
-
-"""
-The Probabilists' version of the Hermite polynomials
-"""
-@memoize function He(x, n::Int)
-    if n == 0
-        return length(x) > 1 ? ones(length(x)) : 1
-    elseif n == 1
-        return x
-    else
-        return x .* He(x, n-1) .- (n-1) .* He(x, n-2)
-    end
-end
-
-
-"""
-The Physicists' version of the Hermite polynomials
-"""
-function H(x, n::Int)
-    return 2^(n/2) * He(x*√2, n)
-end
-
-
-"""
-Convert a standard normal variable to any given marginal distribution
-"""
-function z2x(margin, x)
-    quantile.(margin, cdf.(Normal(0,1), x))
-end
-
+include("common.jl")
 
 function get_coefs(margin, n)
     c = Array{Float64, 1}(undef, n+1)
